Project 1- Flower Example
This is a project assigned in the beginning of the semester, just to refresh our minds of using Java (after the summer break). 

The Flower class gets and sets flower names, count, and price as defined by the FlowerTest class. These are then set up in an array and printed out in order.